1.0.45.36: Fix solaris x86-64 page size issue
     * Patch lp#720800 from Jim Wise

+++ src/runtime/sunos-os.h
@@ -20,7 +20,7 @@
 
 /* FIXME: Stolen from CMUCL. Investigate. */
 typedef unsigned long os_vm_address_t;
-typedef long os_vm_size_t;
+typedef size_t os_vm_size_t;
 typedef off_t os_vm_offset_t;
 typedef int os_vm_prot_t;
